requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# NLP
nltk>=3.8.0
sentence-transformers>=2.2.0

# Analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Progress bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0
